requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2
idna==2.8
python-decouple==3.1
pytz==2018.9
requests==2.21.0
sqlparse==0.3.0
